All,
Our prior public CWB and MetaDataServer is due to retire and its replacement has had to move to the Denver Federal Center from the NEIC.  This means the IP address has changed.

The latest versions of "mdget" and "CWBQuery" at the ftp://hazards.cr.usgs.gov/CWBQuery directory have had their "default" IP address changed to the right place.  Most of you do not use mdget, but if you do, please get the latest source from the FTP area and compile it on your computers.  All of you should update your CWBQuery jar files with the latest ones if you have not done so in the last couple of months.

.....

Hi

Will the domain name remain as cwb-pub.cr.usgs.gov, or will it need to change as well?

In the CWB support in seisFile, I have cwb-pub.cr.usgs.gov as the default server by name an not by IP, so if that stays the same then it will continue to work. If it has changed then I need to put out a new release. Not a big deal either way, just wanted to make sure.

thanks,
Philip

------------------


Phillip,

The cwb-pub address will point to the old one until its retirement.  I may be able to get a alias to the new server done at that time.  But you know how aliases go.

The DNS name for the new server is cwbpub.cr.usgs.gov.  I would change it so that it is right in your next release.

Dave